+++ extensions/cast/src/main/java/com/google/android/exoplayer2/ext/cast/CastPlayer.java
@@ -33,7 +33,6 @@ import com.google.android.exoplayer2.source.TrackGroupArray;
 import com.google.android.exoplayer2.trackselection.FixedTrackSelection;
 import com.google.android.exoplayer2.trackselection.TrackSelection;
 import com.google.android.exoplayer2.trackselection.TrackSelectionArray;
-import com.google.android.exoplayer2.trackselection.TrackSelector;
 import com.google.android.exoplayer2.util.Assertions;
 import com.google.android.exoplayer2.util.Clock;
 import com.google.android.exoplayer2.util.ListenerSet;
@@ -504,12 +503,6 @@ public final class CastPlayer extends BasePlayer {
       default:
         throw new IndexOutOfBoundsException();
     }
-  }
-
-  @Override
-  @Nullable
-  public TrackSelector getTrackSelector() {
-    return null;
   }
 
   @Override